Remove unnecessary code
The removal is necessary because is no longer needed because the update and delete logic is handle directly in the `repository` and not in the model. Also, because we update one or more than one setting at the time, we can quickly trigger the N+1 problem, or trying to rehydrate the model to check if is locked is not necessary. In the case of a deletion we just run a single extra query, and for updating, we get the information directly from the current query, so this does not add any heavy calculation to the code.

Signed-off-by: asciito <ayax.cordova@aydev.mx>